Move implementation to module (2/x)
